slotupdates:  Instead of having to send the full list, a single slot 	      can be sent in a slot update.
Adjusters are now working, but not repositioning the sitter.  Probably
doSeats() needs to be called to position the sitter every time the
slot is updated (i. e. every time the adjuster is moved, and when
adjusting is stopped).

The core has a bug or glitch in that it somehow ties the seat a sitter
is put into to the slot number rather than to the seat number.  This
means that when an adjuster is used and the slot number the seat is in
changes due to updating the slot (i. e. seat gets a different stride
in the slots list), the sitter may be put into an unexpected seat when
they sit again.

Besides, adjusting has the problem that on every switch of a pose, the
notecard is read again and the adjustment gets overwritten.  This
makes adjusting even more awkward than it genuinely is anyway.

Memory consumption needs to be fixed.